backend/requirements.txt
langgraph
langchain
langchain-community
openai
faiss-cpu
pypdf
fastapi
uvicorn
python-dotenv
python-multipart
